# openapi.json
{"openapi":"3.0.1","info":{"title":"Spring Data I","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Review APIs","description":"APIs for managing reviews"},{"name":"Address APIs","description":"APIs for managing addresses"},{"name":"Task 3 APIs","description":"Implement the following queries by using Naming Convention. Create REST endpoints for them."},{"name":"Product APIs","description":"APIs for managing products"},{"name":"User APIs","description":"APIs for managing users"},{"name":"Category APIs","description":"APIs for managing categories"}],"paths":{"/user":{"put":{"tags":["User APIs"],"summary":"Update an existing user","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"post":{"tags":["User APIs"],"summary":"Create a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/review":{"put":{"tags":["Review APIs"],"summary":"Update an existing review","operationId":"updateReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}},"post":{"tags":["Review APIs"],"summary":"Create a new review","operationId":"createReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}}},"/product":{"put":{"tags":["Product APIs"],"summary":"Update an existing product","operationId":"updateProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"post":{"tags":["Product APIs"],"summary":"Create a new product","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/category":{"put":{"tags":["Category APIs"],"summary":"Update an existing category","operationId":"updateCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}},"post":{"tags":["Category APIs"],"summary":"Create a new category","operationId":"createCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}}},"/address":{"put":{"tags":["Address APIs"],"summary":"Update an existing address","operationId":"updateAddress","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Address"}}}}}},"post":{"tags":["Address APIs"],"summary":"Create a new address","operationId":"createAddress","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Address"}}}}}}},"/user/{id}":{"get":{"tags":["User APIs"],"summary":"Get a user by Id","operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["User APIs"],"summary":"Delete a user by Id","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/review/{id}":{"get":{"tags":["Review APIs"],"summary":"Get a review by Id","operationId":"getReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}},"delete":{"tags":["Review APIs"],"summary":"Delete a review by Id","operationId":"deleteReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/review/product/{productId}":{"get":{"tags":["Review APIs","Task 3 APIs"],"summary":"Get reviews by product","operationId":"getReviewsByProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int64"},"example":4}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"}}}}}}}},"/product/{id}":{"get":{"tags":["Product APIs"],"summary":"Get a product by Id","operationId":"getProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["Product APIs"],"summary":"Delete a product by Id","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/product/price":{"get":{"tags":["Task 3 APIs","Product APIs"],"summary":"Find all products that cost more than minPrice","operationId":"getProductsByPriceGreaterThan","parameters":[{"name":"minPrice","in":"query","required":true,"schema":{"type":"number"},"example":200}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/product/name":{"get":{"tags":["Task 3 APIs","Product APIs"],"summary":"Find all products that contain keyword in the name.","operationId":"getProductsByNameContaining","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"},"example":"Shorthair"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/product/category":{"get":{"tags":["Task 3 APIs","Product APIs"],"summary":"Find all products in cat category and cost less than maxPrice","operationId":"getProductsByPriceLessThanAndCategoryIs","parameters":[{"name":"maxPrice","in":"query","required":true,"schema":{"type":"number"},"example":3000},{"name":"categoryId","in":"query","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/category/{id}":{"get":{"tags":["Category APIs"],"summary":"Get a category by Id","operationId":"getCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}},"delete":{"tags":["Category APIs"],"summary":"Delete a category by Id","operationId":"deleteCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/address/{id}":{"get":{"tags":["Address APIs"],"summary":"Get an address by Id","operationId":"getAddress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Address"}}}}}},"delete":{"tags":["Address APIs"],"summary":"Delete an address by Id","operationId":"deleteAddress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Address":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"street":{"type":"string"},"zip":{"type":"string"},"city":{"type":"string"}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"},"rating":{"type":"string"},"category":{"$ref":"#/components/schemas/Category"},"reviews":{"type":"array","items":{"$ref":"#/components/schemas/Review"}}}},"Review":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"comment":{"type":"string"},"reviewedProduct":{"$ref":"#/components/schemas/Product"},"submittedUser":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"submittedReviews":{"type":"array","items":{"$ref":"#/components/schemas/Review"}},"address":{"$ref":"#/components/schemas/Address"}}}}}}